Remove code, change code
Remove some now-unnecessary code for determining if a line contains a
heading, as this is now being done earlier in the code.
Change code to no longer count the total number of number signs on a
line, as only the number signs at the beginning and end of the line are
relevant.